requirements.txt
pytorch>=1.7.1
torchvision
opencv-python
transforms3d
numpy
plyfile
scikit-image
matplotlib
open3d
pytorch3d
tensorboardX
tqdm
